Use a less resource-intensive notify library
This allows you to use lrt on a mac without having to increase the
number of open file-descriptors. As a result we can also be more precise
about file-edits, and will track edits to any code included by your
program.

It is slightly slower (I had to set the timeout to 200ms from 100ms to
make the tests pass) but that should still be within the realm of
reasonable for making new web-requests.